ticker snapshot

+++ src/main/java/org/capitalcompass/capitalcompassstocks/model/TickerSnapshot.java
@@ -2,12 +2,10 @@ package org.capitalcompass.capitalcompassstocks.model;
 
 import lombok.Data;
 
-import javax.validation.constraints.Min;
-
 @Data
 public class TickerSnapshot {
 
-    private Integer updated;
+    private Long updated;
     private DailyBar day;
     private DailyBar prevDay;
 }
